refactor: Add header comments to each .c file and separate dispatchable functions
