+++ src/tasks/cosmic_monkey/cosmic_monkey_helpers.c
@@ -1,6 +1,24 @@
+/**
+ * cosmic_monkey_helpers.c
+ * 
+ * Helper functions for the Cosmic Monkey task. This is an antagonist task loosely inspired by
+ * Netflix's Chaos Monkey. It randomly flips bits in memory to create an environment akin to low-earth 
+ * orbit in order to test the robustness of OS logic to cosmic radiation.
+ * 
+ * Created: October 6, 2024
+ * Authors: Oren Kohavi, Ignacio Blancas Rodriguez
+ */
+
 #include "atmel_start.h"
 #include "cosmic_monkey_task.h"
 #include "logging.h"
+
+/* ---------- DISPATCHABLE FUNCTIONS (sent as commands through the command dispatcher task) ---------- */
+
+// NOTE: The Cosmic Monkey task is an antagonist task that won't be run in release builds. It has no
+// command queue, nor will it be able to receive commands from the command dispatcher task.
+
+/* ---------- NON-DISPATCHABLE FUNCTIONS (do not go through the command dispatcher) ---------- */
 
 static const int VALID_MEMORY_RANGE_START = 0x20000000;
 static const int VALID_MEMORY_RANGE_IN_BYTES = 18;
