+++ src/tasks/radio/radio_task.c
@@ -1,0 +1,12 @@
+/**
+ * radio_task.c
+ * 
+ * ...
+ * 
+ * Created: ...
+ * Authors: ...
+ */
+
+/* ---------- DISPATCHABLE FUNCTIONS (sent as commands through the command dispatcher task) ---------- */
+
+/* ---------- NON-DISPATCHABLE FUNCTIONS (do not go through the command dispatcher) ---------- */
