+++ src/tasks/task_manager/task_manager_main.c
@@ -1,3 +1,14 @@
+/**
+ * task_manager_helpers.c
+ * 
+ * Helper functions for the task manager task. This task is responsible for initializing and enabling/disabling
+ * all other tasks in the system based on PVDX's state diagram.
+ * 
+ * Created: April 14, 2024
+ * Authors: Oren Kohavi, Ignacio Blancas Rodriguez, Tanish Makadia, Yi Liu, Siddharta Laloux, Aidan Wang, Simon Juknelis,
+ * Defne Doken, Aidan Wang, Jai Garg, Alex Khosrowshahi
+ */
+
 #include "task_manager_task.h"
 
 task_manager_task_memory_t task_manager_mem;
@@ -7,35 +18,31 @@ SemaphoreHandle_t task_list_mutex = NULL;
 StaticSemaphore_t task_list_mutex_buffer;
 
 void main_task_manager(void *pvParameters) {
-    info("task-manager: Started main loop\n");
+    info("task-manager: Task Started!\n");
 
     // Enqueue a command to initialize all subtasks
-    status_t result;
-    command_t command_task_manager_init_subtasks = {TASK_MANAGER, OPERATION_INIT_SUBTASKS, NULL, 0, &result, NULL};
+    command_t command_task_manager_init_subtasks = {TASK_MANAGER, OPERATION_INIT_SUBTASKS, NULL, 0, NULL, NULL};
     enqueue_command(&command_task_manager_init_subtasks);
+
+    // Cache the watchdog checkin command to avoid creating it every iteration
+    command_t cmd_checkin = get_watchdog_checkin_command();
 
     command_t cmd;
     BaseType_t xStatus;
 
-    TaskHandle_t handle = xTaskGetCurrentTaskHandle();
-    command_t command_checkin = {TASK_WATCHDOG, OPERATION_CHECKIN, &handle, sizeof(TaskHandle_t*), NULL, NULL};
-
     while (true) {
-        debug("task_manager: Started main loop\n");
-        // if there's something in the queue, pop it and execute it
-        xStatus = xQueueReceive(task_manager_command_queue_handle, &cmd, 0);
+        debug("\n---------- Task Manager Task Loop ----------\n");
 
-        if (xStatus == pdPASS) {
-            // Command received, so execute it
+        // Dispatch all commands contained in the queue
+        while ((xStatus = xQueueReceive(task_manager_command_queue_handle, &cmd, 0)) == pdPASS) {
             debug("task_manager: Command popped off queue. Target: %d, Operation: %d\n", cmd.target, cmd.operation);
             exec_command_task_manager(cmd);
         }
-        else {
-            // No command received, so continue
-            debug("task_manager: No commands queued.\n");
-        }
+        debug("task_manager: No more commands queued.\n");
         
-        enqueue_command(&command_checkin);
+        // Check in with the watchdog task
+        enqueue_command(&cmd_checkin);
+        // Wait 1 second before attempting to run the loop again
         vTaskDelay(pdMS_TO_TICKS(1000));
     }
 }
